Adamm

+++ Student/student_application_list.php
@@ -30,35 +30,56 @@
 
             <tbody>
                 <?php
-                include("../config.php");
-                session_start();    
-                // Retrieve data from table
-                $id = $_SESSION['USER_ID'];
-                $query = "SELECT * FROM practical_training WHERE fk_userid = '$id'";
-                $sql = mysqli_query($conn, $query) or die(mysqli_connect_error());
-                
-                if (mysqli_num_rows($sql) > 0) {
-                    // Output data of each row
-                    while ($row = mysqli_fetch_array($sql)) {
-                        echo "
-                        <tr>
-                            <td>$row[applicationid]</td>
-                            <td>$row[applicationdate]</td>
-                            <td>$row[applicationtitle]</td>
-                            <td>$row[applicationstatus]</td>
-                            <td>
-                                <a class='btn btn-primary btn-sm' href='edit_user_form.php?id=$row[applicationid]'>Edit</a>
-                                <a class='btn btn-danger btn-sm' href='delete_user.php?id=$row[applicationid]'>Delete</a>
-                                <a class='btn btn-dark btn-sm' href='view_user.php?id=$row[applicationid]'>View</a>
-                            </td>
-                        </tr> 
-                        ";
-                    }
-                } else {
-                    echo "0 results";
-                }
 
-                mysqli_close($conn);
+                session_start();
+
+                // Retrieve data for the logged-in student
+                $studentId = $_SESSION['USER_ID']; 
+
+                  // Check if the student ID is set
+                  if (isset($studentId)) {
+                    require_once("../config.php");
+                    require_once("../functions.php");
+
+                    // Retrieve data from the table
+                    $array = array();
+                    $select = "SELECT * FROM practical_training WHERE fk_userid = $studentId";
+                    $sql = mysqli_query($GLOBALS['conn'], $select);
+
+                    // Check if the query executed successfully
+                    if ($sql) {
+                        if (mysqli_num_rows($sql) > 0) {
+                            // Output data of each row
+                            while ($row = mysqli_fetch_array($sql)) {
+                              $array['userid'] = $row['fk_userid'];
+                              $profile = getUsersData($array['userid']);
+                              echo "
+                              <tr>
+                                  <td>$row[applicationid]</td>
+                                  <td>$profile[name]</td>
+                                  <td>$row[applicationdate]</td>  
+                                  <td>$row[applicationtitle]</td>
+                                  <td>$row[applicationstatus]</td>
+                                  <td>
+                                      <a class='btn btn-primary btn-sm' href='edit_user_form.php?id=$row[applicationid]'>Edit</a>
+                                      <a class='btn btn-danger btn-sm' href='delete_user.php?id=$row[applicationid]'>Delete</a>
+                                      <a class='btn btn-dark btn-sm' href='view_user.php?id=$row[applicationid]'>View</a>
+                                  </td>
+                              </tr> 
+                              ";
+                            }
+                        } else {
+                            echo "0 results";
+                        }
+
+                        mysqli_close($conn);
+                    } else {
+                        echo "Query execution failed.";
+                    }
+                  } else {
+                    echo "Student ID is not set.";
+                  }
+
                 ?>
             </tbody>
         </table>
